Add logic to show different messages if you haven't walked yet today
