+++ client/src/components/Message/Message.scss
@@ -1,0 +1,15 @@
+@use '../../styles/partials/variables' as *;
+@use '../../styles/partials/mixins' as *;
+
+.message {
+    color: #ffffff;
+    width: 75%;
+    text-align: center;
+    position: relative;
+    top: 2rem;
+
+    @include tablet {
+       top: 4rem;
+       font-size: 1.25rem;
+    }
+}
