+++ client/src/pages/Home/Home.js
@@ -1,23 +1,25 @@
-import BottomNav from '../../components/BottomNav/BottomNav';
-import AnimatedSittingDog from '../../components/AnimatedSittingDog/AnimatedSittingDog';
 import TopNav from '../../components/TopNav/TopNav';
+import BottomNav from '../../components/BottomNav/BottomNav';
+import Message from '../../components/Message/Message';
+import AnimatedSittingDog from '../../components/AnimatedSittingDog/AnimatedSittingDog';
 import HomeCard from '../../components/HomeCard/HomeCard';
-import { getRandomElement } from '../../utils/mathUtils';
+import { useState, useEffect } from 'react';
+import { getTodaysSteps } from '../../utils/storageUtils';
 import './Home.scss';
+
 
 function Home({ signedIn }) {
 
-    let happyMessages = [
-        "Have you walked yet today?",
-        "Take 5 minutes to stretch",
-        "Paws for thought: a walk a day keeps the stress away",
-        "Every step is a step toward a happier you and a happier dog",
-        "Take a break and grab a cup of water",
-    ];
-    let sadMessages = [
-        "Stretch those legs and let your dog stretch its paws!", 
-        "Your dog needs a walk - take some time to move",
-    ];
+    let [ steps, setSteps ] = useState({});
+
+    useEffect(() => {
+        getTodaysSteps(
+            signedIn,
+            (data) => {
+                data.length === 0 ? setSteps({steps: 0}) : setSteps(data[0])}, 
+            (error) => console.log('Error getting steps', error)
+        );
+    }, []);
 
     return (
         <div className='page page--home'>
@@ -25,13 +27,13 @@ function Home({ signedIn }) {
                 <div className="nav__container">
                     <TopNav page='home' signedIn={signedIn} />
                 </div>
-                <p className="message">{getRandomElement(happyMessages)}</p>
+                <Message steps={steps} />
                 <AnimatedSittingDog />
                 <BottomNav page='home' />
             </div>
             <div className="page__right">
                 <TopNav page='home' signedIn={signedIn} />
-                <HomeCard signedIn={signedIn} />
+                <HomeCard steps={steps} />
             </div>
         </div>
     );
